better-test-code/src/main/java/org/example/better/CatanResourceCalculator.java
package org.example.better;

import java.util.ArrayList;
import java.util.List;
import java.util.Map;

public class CatanResourceCalculator {

    public static void main(String[] args) {
        List<Player> players = initializePlayers();
        calculateAndPrintSettlementPossibilities(players);
    }
    
    public static List<Player> initializePlayers() {
        return new ArrayList<>();
    }

    public static void calculateAndPrintSettlementPossibilities(List<Player> players) {
        for (Player player : players) {
            if (hasEnoughResourcesForSettlement(player)) {
                List<Trade> possibleTrades = calculatePossibleTrades(player);
                if (canBuildSettlementAfterTrades(player, possibleTrades)) {
                    System.out.println("Player " + player.getName() + " can build a settlement!");
                }
            }
        }
    }

    // Helper methods (now more focused and testable)
    public static boolean hasEnoughResourcesForSettlement(Player player) {
        Map<ResourceType, Integer> resources = player.getResources();
        return resources.containsKey(ResourceType.BRICK) && resources.get(ResourceType.BRICK) > 0 && 
            resources.containsKey(ResourceType.SHEEP) && resources.get(ResourceType.SHEEP) > 0 && 
            resources.containsKey(ResourceType.WHEAT) && resources.get(ResourceType.WHEAT) > 0 && 
            resources.containsKey(ResourceType.WOOD) && resources.get(ResourceType.WOOD) > 0;
    }

    private static List<Trade> calculatePossibleTrades(Player player) {
        // Calculate all possible trades...
        // ...
        // TODO implement
        return null;
    }

    private static boolean canBuildSettlementAfterTrades(Player player, List<Trade> trades) {
        // Check if the player can build a settlement after applying trades...
        // (more complex logic)
        // ...
        // TODO implement
        return false;
    }

    // Other methods (initializePlayers, etc.) not shown...
}